refactor: Line와 LineMaker 객체 통합
